+++ domain/src/main/java/com/scccrt/klima/domain/repository/WeatherRepository.kt
@@ -4,5 +4,5 @@ import com.scccrt.klima.domain.model.CurrentWeather
 
 interface WeatherRepository {
 
-    suspend fun getWeather(lat: Long, long: Long): Result<CurrentWeather>
+    suspend fun getWeather(lat: Double, long: Double): Result<CurrentWeather>
 }
